Add files via upload
Deepfakes is a term used to refer to a video or image that has been edited to replace the
person in the video with someone else using an algorithm. Deep fake algorithms can create
fake images and videos that humans cannot distinguish from authentic ones.
The term ‘Deep Fake’ comes from ‘Deep Learning’ where the machines teach themselves how
to solve problems given large sets of data. Deepfakes are usually created using deep neural
networks that basically perform face-swapping. This technique that is used by deep fakes is so
much more powerful than image-to-image translation as it doesn’t require training data.
Machine Learning is added on to this in the form of GANs or (Generative Adversarial Networks)
which on the basis of the study of large amounts of data detects and improves any flaws in the
deep fake within multiple rounds, making a deep fake almost undetectable to the human eye.
By reviewing the background of deep fakes, this study provides a consolidated review of deep
fake techniques and tries to implement a more robust method to deal with the increasingly
challenging problem of deep fakes. The aim of this paper is to implement an algorithm that
basically reverses engineers the generative algorithm to detect a deep fake from a genuine
video or an image.